Add use-toggle-resource-embedding hook
